lua: reorganize `api/` so each file is one API namespace
Renamed `agent_api.rs`, `async_api.rs`, `fn_api.rs` to drop the `_api` suffix (using `r#async` / `r#fn` raw identifiers for the keyword conflicts).

Moved `buf.rs` and `win.rs` under `ui/` since they belong to that namespace, and moved `command.rs`, `ctx.rs`, `setup.rs` into a new `util/` subdirectory. The `err_pair`, `json_to_lua`, `lua_to_json` helpers that lived in `mod.rs` now live in `util/convert.rs`.

Now every file directly in `api/` maps 1:1 to a `maki.*` Lua namespace, which makes the directory much easier to navigate.
