+++ maki-lua/src/api/ui/mod.rs
@@ -6,12 +6,15 @@ use std::time::Duration;
 use humantime::format_duration;
 use mlua::{Lua, Result as LuaResult, Table};
 
-use crate::api::command::{
+use crate::api::util::command::{
     Anchor, Border, Dimension, FloatConfig, HintEntries, HintWriter, Split, TitlePos, UiAction,
     WinCommand, WinEvent,
 };
-use crate::api::win::WinHandle;
+pub(crate) mod buf;
+pub(crate) mod win;
+
 use crate::runtime::with_task_bufs;
+use win::WinHandle;
 
 pub(crate) struct HintStore {
     hints: BTreeMap<Arc<str>, Vec<(String, String)>>,
@@ -182,7 +185,7 @@ pub(crate) fn create_ui_table(
             "open_win",
             lua.create_function(
                 move |_lua, (buf_ud, opts_tbl): (mlua::AnyUserData, Table)| {
-                    let buf_handle = buf_ud.borrow::<crate::api::buf::BufHandle>()?;
+                    let buf_handle = buf_ud.borrow::<buf::BufHandle>()?;
                     let title: String = opts_tbl.get("title").unwrap_or_default();
                     let cursor_line: bool = opts_tbl.get("cursor_line").unwrap_or(false);
                     let footer = parse_footer(&opts_tbl)?;
