+++ maki-lua/src/api/ui/win.rs
@@ -1,9 +1,9 @@
 use mlua::{Table, UserData, UserDataMethods};
 
-use crate::api::command::{
+use super::{parse_footer, try_parse_dimension};
+use crate::api::util::command::{
     Anchor, Border, FloatConfigPatch, Split, TitlePos, WinCommand, WinEvent,
 };
-use crate::api::ui::{parse_footer, try_parse_dimension};
 
 pub(crate) struct WinHandle {
     event_rx: flume::Receiver<WinEvent>,
