+++ maki-lua/src/lib.rs
@@ -5,18 +5,18 @@ mod loader;
 pub(crate) mod plugin_permissions;
 mod runtime;
 
-pub use api::command::{
+pub use api::keymap::{KeymapEntry, KeymapReader, KeymapSnapshot};
+pub use api::util::command::{
     Anchor, Axis, Border, Dimension, Edge, FloatConfig, FloatConfigPatch, HintReader, HintSnapshot,
     LuaCommandInfo, LuaCommandReader, Split, TitlePos, UiAction, WinCommand, WinEvent,
 };
-pub use api::keymap::{KeymapEntry, KeymapReader, KeymapSnapshot};
 pub use error::PluginError;
 pub use loader::{EventHandle, PluginHost};
 pub use plugin_permissions::{Permission, PluginPermissions};
 pub use runtime::{ClickReply, RestoreItem};
 
 pub mod test_support {
-    use crate::api::command::{LuaCommandInfo, LuaCommandReader, LuaCommandWriter};
+    use crate::api::util::command::{LuaCommandInfo, LuaCommandReader, LuaCommandWriter};
 
     pub struct LuaCommandWriterHandle(LuaCommandWriter);
